Design Refactoring
