+++ cas-client-core/src/main/java/org/apereo/cas/client/ssl/WhitelistHostnameVerifier.java
@@ -18,7 +18,6 @@
  */
 package org.apereo.cas.client.ssl;
 
-import javax.net.ssl.HostnameVerifier;
 import javax.net.ssl.SSLSession;
 import java.io.Serializable;
 
@@ -30,7 +29,7 @@ import java.io.Serializable;
  * @since 3.1.10
  *
  */
-public final class WhitelistHostnameVerifier implements HostnameVerifier, Serializable {
+public final class WhitelistHostnameVerifier extends AbstractHostnameVerifier implements Serializable {
 
     private static final long serialVersionUID = 1L;
