+++ cas-client-core/src/main/java/org/apereo/cas/client/util/CloseableUtil.java
@@ -1,0 +1,17 @@
+package org.apereo.cas.client.util;
+
+import java.io.Closeable;
+import java.io.IOException;
+
+public class CloseableUtil {
+
+    public static void closeQuietly(final Closeable resource) {
+        try {
+            if (resource != null) {
+                resource.close();
+            }
+        } catch (final IOException e) {
+            //ignore
+        }
+    }
+}
